requirements.txt
streamlit
gtts
speechrecognition
google-generativeai
pyaudio ; platform_system == "Windows"
